# server/src/docs/swagger.json
{
	"definitions": {
		"CartI": {
			"description": "",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"userId": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"productId": {
					"type": "integer",
					"format": "int32",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"userId",
				"productId"
			]
		},
		"UserI": {
			"description": "",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"lastName": {
					"type": "string",
					"description": ""
				},
				"email": {
					"type": "string",
					"description": ""
				},
				"password": {
					"type": "string",
					"description": ""
				},
				"idDocumentType": {
					"type": "string",
					"description": ""
				},
				"idDocumentNumber": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"rol": {
					"type": "string",
					"description": ""
				},
				"status": {
					"type": "integer",
					"format": "int32",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"lastName",
				"email",
				"idDocumentType",
				"idDocumentNumber",
				"rol",
				"status"
			]
		},
		"ProductI": {
			"description": "",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				},
				"price": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"discountPercentage": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"currentStock": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"reorderPoint": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"minimum": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"storeId": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"categories": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"sizes": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"brand": {
					"type": "string",
					"description": ""
				},
				"url_img": {
					"type": "string",
					"description": ""
				},
				"status": {
					"type": "integer",
					"format": "int32",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"description",
				"price",
				"discountPercentage",
				"currentStock",
				"reorderPoint",
				"minimum",
				"storeId",
				"categories",
				"sizes",
				"brand",
				"url_img"
			]
		},
...
	},
	"info": {
		"description": "Shoppy API.",
		"license": {
			"name": "ISC"
		},
		"title": "shoppy-api",
		"version": "1.0.0"
	},
	"paths": {
		"/api/cart/": {
			"post": {
				"operationId": "CartGet",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Retrieve a cart.",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/CartI"
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "Cart not found."
					}
				},
				"description": "GET a cart",
				"tags": [
					"Cart"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"decodedToken": {
									"$ref": "#/definitions/UserI"
								}
							}
						}
					}
				]
			},
			"delete": {
				"operationId": "CartDelete",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"Cart"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"entity": {
									"$ref": "#/definitions/ProductI"
								},
								"decodedToken": {
									"$ref": "#/definitions/UserI"
								}
							}
						}
					}
				]
			}
		},
		"/api/cart/add": {
			"post": {
				"operationId": "CartAdd",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"Cart"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"entity": {
									"$ref": "#/definitions/ProductI"
								},
								"decodedToken": {
									"$ref": "#/definitions/UserI"
								}
							}
						}
					}
				]
			}
		},
		"/api/cart/deleteAll": {
			"delete": {
				"operationId": "CartDeleteAll",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"Cart"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"decodedToken": {
									"$ref": "#/definitions/UserI"
								}
							}
						}
					}
				]
			}
		},
		"/api/product/{productId}": {
			"get": {
				"operationId": "ProductGetById",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"201": {
						"description": "Retrieve a Product.",
						"schema": {
							"$ref": "#/definitions/ProductI"
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "Product not found."
					}
				},
				"description": "GET a specific product from it's id",
				"tags": [
					"Products"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "productId",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			}
		},
		"/api/product/": {
			"get": {
				"operationId": "ProductGetAll",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"201": {
						"description": "Retrieve a list of Products.",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/ProductI"
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "Products not found."
					}
				},
				"description": "GET all products from a store id",
				"tags": [
					"Products"
				],
				"parameters": []
			},
			"post": {
				"operationId": "ProductPost",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"201": {
						"description": "Insert a Product on the Database.",
						"schema": {
							"$ref": "#/definitions/ProductI"
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "Product insert failed."
					}
				},
				"description": "CREATE a product on the database",
				"tags": [
					"Products"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "product",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ProductI"
						}
					}
				]
			},
			"put": {
				"operationId": "ProductUpdate",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Update a Product on the Database. Except Categories and Sizes",
						"schema": {
							"$ref": "#/definitions/ProductI"
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "Product not found."
					}
				},
				"description": "UPDATE a product",
				"tags": [
					"Products"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "product",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ProductI"
						}
					}
				]
			}
		},
		"/api/product/{id}": {
			"delete": {
				"operationId": "ProductDelete",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Delete a Product on the Database.",
						"schema": {
							"$ref": "#/definitions/ProductI"
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "Product not found."
					}
				},
				"description": "",
				"tags": [
					"Products"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			}
		},
...
	},
	"swagger": "2.0",
	"securityDefinitions": {
		"client secret": {
			"type": "apiKey",
			"name": "x-auth",
			"in": "header"
		}
	}
}